Patch to fix IB collection in systems with multiple mellanox cards suggested-by: Gary Skouson
The use of HCAName isn't right and should be HCAId, especially if you have mellanox ethernet and IB in the same box, the '0' or $i isn't going to work right. For example on one system, we have bonded ethernet on a mellanox ethernet card and a separate IB card, so /sys/class/infiniband shows two devices mlx5_2 and mlx5_bond_0.

For the formatit.ph the logic isn't right. The $link comparison should be with "eq" not "=". I also think we need to chomp $link to remove the newline before the compare.